refactor(hass,sensors): replace interface{} usage with json.RawMessage
Start replacing usage of interface{} with json.RawMessage. Most cases
where interface{} is being used, is just a situation where an unknown
JSON structure exists. While it is unknown at that point, it is in an
intermediate capacity and the actual structure is not needed. So using
json.RawMessage provides the advantage of being more explicit and
deferring the marshalling till when it needs to happen.