feat(linux): add a Data Source attribute to all Linux sensors
The "Data Source" reflects where the data for the sensor came from. For
example, if the data was sourced from D-Bus, it will have the value of
"D-Bus". If the data was sourced from procfs, it will be "procfs".

This might be useful for grouping sensors in the Home Assistant UI
and/or in templating/scripts.