feat(tracker)!: migration of registry to new format
The registry used by the tracker for keeping state of the
registered/disabled status of a sensor has been migrated from nutsdb to
json files. In doing this change, the registry is now simpler to manage
and saves significant memory. The registry should be automatically
migrated through the agent upgrade function.

Minor additional change, the IsRegistered/Disabled functions for the
Registry interface now signal via a channel instead of passing the value
directly.