refactor(linux): migrate sensors to common struct
This significant refactor is the first step to moving all Linux sensors
to use a common struct as their base, where possible. This common struct
satisfies the tracker.Sensor interface so it can be used as a sesnor
update. This change removes a large amount of duplicated code for each
sensor that was required to satisfy that interface.