refactor!: major MQTT functionality refactor
BREAKING CHANGE: This commit is a major
refactoring of the MQTT functionality coinciding
with changes to the underlying library that powers
it. **MQTT entities have been renamed, which will
result in some breakage of automations and
features in Home Assistant.** No functionality has
been lost however, and this change will make it
easier to add additional controls and features
powered by MQTT to Go Hass Agent.